requirements.txt
djangorestframework
django
coverage
drf-yasg
setuptools
